backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
fastapi-socketio
python-multipart
python-dotenv
psycopg2-binary
slowapi
